+++ CHANGELOG.md
@@ -3,8 +3,12 @@
 Releases in reverse chronological order.
 
 Please check
-[GitHub](https://github.com/micro-os-plus/devices-stm32f0-xpack/issues/)
+[GitHub](https://github.com/micro-os-plus/devices-stm32f0-extras-xpack/issues/)
 and close existing issues and pull requests.
+
+## 2021-01-15
+
+- rename as `devices-stm32f0-extras`
 
 ## v1.1.0 (2021-02-04)